Use -no-undefined linker flag for building salut plugin
So it builds as a DLL on windows

NOTE: this doesn't make the linker whine about undefined symbols on linux. This is a libtool policy
libtool -no-undefined is a no-op on Linux. So we need the buildbot windows target to verify we don't
end up using symbols which aren't really in the salut-plugins library.