*/status.c: don't crash if form fields have unexpected types
FORM_TYPE should be hidden, type should be text-single, and name and caps
should be text-multi.

This is significant because if the types of name and caps are not
explicitly specified in the message, Wocky will guess. If there is only
one value, it will guess text-single, and the GValue will contain
G_TYPE_STRING, not G_TYPE_STRV, causing a crash when we call
g_value_get_boxed.

Reviewed-by: Olli Salli <olli.salli@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=45236